Request View Refactoring
