+++ src/react/src/components/RequestItemCard.tsx
@@ -12,15 +12,13 @@ function RequestItemCard({
   listeners,
   updateRequestItem,
   removeItem,
-  addItem,
   isDialog,
   config,
 }: {
   requestItem: RequestItem;
   listeners: Record<string, any>;
-  updateRequestItem: (item: RequestItem) => void;
+  updateRequestItem: (itemParams?: Partial<RequestItem>) => void;
   removeItem: (id: string) => void;
-  addItem: (itemParams?: Partial<RequestItem>) => void;
   isDialog: boolean;
   config: Config;
 }) {
@@ -54,7 +52,6 @@ function RequestItemCard({
           updateRequestItem={updateRequestItem}
           removeItem={removeItem}
           listeners={listeners}
-          addItem={addItem}
           isDialog={isDialog}
           config={config}
         />
