+++ src/react/src/components/RequestOptions.tsx
@@ -1,0 +1,223 @@
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Dropdown } from "primereact/dropdown";
+import { MenuItem } from "primereact/menuitem";
+import { SplitButton } from "primereact/splitbutton";
+import { useNavigate } from "react-router-dom";
+
+import AccessButton from "../components/AccessButton";
+import { Request } from "../models/brewtils-types";
+import {
+  Config,
+  PermissionCheck,
+  RequestCommand,
+  RequestItem,
+} from "../models/models";
+import { checkPermission } from "../services/permission_service";
+import { CancelRequest, DeleteRequest } from "../services/request_service";
+import { GetBaseURL } from "../services/util_service";
+
+const handleDownload = (request: Request) => {
+  // Example: fetch a file from a URL
+  const fileUrl = `${GetBaseURL()}/api/v1/requests/output/${request.id}`;
+  let filename = `${request.id}.txt`;
+  if (request.output_type == "HTML") {
+    filename = `${request.id}.html`;
+  } else if (request.output_type == "JSON") {
+    filename = `${request.id}.json`;
+  }
+
+  fetch(fileUrl)
+    .then((response) => response.blob())
+    .then((blob) => {
+      // Create blob link to download
+      const url = window.URL.createObjectURL(new Blob([blob]));
+      const link = document.createElement("a");
+      link.href = url;
+      link.setAttribute("download", filename); // Set the custom download name
+      document.body.appendChild(link);
+      link.click(); // Trigger the download
+      link?.parentNode?.removeChild(link); // Clean up the link
+      window.URL.revokeObjectURL(url); // Free up the memory
+    })
+    .catch((error) => {
+      console.error("Error fetching the file:", error);
+    });
+};
+
+function RequestOptions({
+  request,
+  setRequest,
+  requestProjections,
+  requestProjectionSelected,
+  setRequestProjectionSelected,
+  requestProjectionSelectedRef,
+  addRequestItem,
+  config,
+  isCard,
+  openRequest,
+}: {
+  request: Request;
+  setRequest: (request: Request | null) => void;
+  requestProjections?: RequestCommand[];
+  requestProjectionSelected?: RequestCommand;
+  setRequestProjectionSelected: (value: RequestCommand | undefined) => void;
+  requestProjectionSelectedRef: React.RefObject<RequestCommand | undefined>;
+  addRequestItem: (itemParams?: Partial<RequestItem>) => void;
+  config: Config;
+  isCard: boolean;
+  openRequest?: () => void;
+}) {
+  const navigate = useNavigate();
+  const items: MenuItem[] = [];
+
+  const execute_authority = checkPermission(config, "OPERATOR", {
+    gardenName: request?.target_garden,
+  } as PermissionCheck);
+
+  if (execute_authority) {
+    if (
+      request.status &&
+      ["CREATED", "RECEIVED", "IN_PROGRESS"].includes(request.status)
+    ) {
+      items.push({
+        label: "Cancel Request",
+        icon: <FontAwesomeIcon icon="xmark" />,
+        command: () => {
+          CancelRequest(request).catch((error) => {
+            console.error("Error canceling request:", error);
+          });
+        },
+      });
+    } else {
+      if (isCard) {
+        items.push({
+          label: "Pour Again",
+          icon: <FontAwesomeIcon icon="plus" />,
+          command: () => {
+            pourAgain(request);
+          },
+        });
+      }
+      items.push({
+        label: "Download Output",
+        icon: <FontAwesomeIcon icon="download" />,
+        command: () => {
+          handleDownload(request);
+        },
+      });
+
+      if (
+        checkPermission(config, "GARDEN_ADMIN", {
+          gardenName: request?.target_garden,
+        } as PermissionCheck)
+      ) {
+        items.push({
+          label: "Delete Request",
+          icon: <FontAwesomeIcon icon="xmark" />,
+          command: () => {
+            DeleteRequest(request)
+              .then(() => {
+                void navigate(`/requests`);
+              })
+              .catch((error) => {
+                console.error("Error deleting request:", error);
+              });
+          },
+        });
+      }
+    }
+    items.push({
+      label: "Reload Request",
+      icon: <FontAwesomeIcon icon="arrows-rotate" />,
+      command: () => {
+        setRequest(null);
+      },
+    });
+  }
+
+  const pourAgain = (request: Request) => {
+    addRequestItem({ requestId: request.id, type: "REQUEST" });
+  };
+
+  const commandTemplate = (requestCommand: RequestCommand) => {
+    return (
+      <span>
+        {requestCommand?.namespace === request.namespace
+          ? null
+          : `${requestCommand?.namespace} / `}{" "}
+        {requestCommand?.systemName} / {requestCommand?.version} /{" "}
+        {requestCommand?.instance} / {requestCommand?.command}
+      </span>
+    );
+  };
+
+  return (
+    <div className="card justify-content-end">
+      <div className="flex flex-end">
+        {execute_authority && (
+          <SplitButton
+            label={isCard ? "Open Request" : "Pour Again"}
+            icon={
+              isCard ? (
+                <FontAwesomeIcon icon="up-right-from-square" className="mr-2" />
+              ) : (
+                <FontAwesomeIcon icon="plus" className="mr-2" />
+              )
+            }
+            model={items}
+            className="p-button-secondary"
+            onClick={() => {
+              if (isCard) {
+                if (openRequest) {
+                  openRequest();
+                }
+              } else {
+                pourAgain(request);
+              }
+            }}
+            severity="success"
+            style={{ marginLeft: "auto" }}
+          />
+        )}
+        {!execute_authority && (
+          <AccessButton
+            icon={<FontAwesomeIcon icon="download" />}
+            label="Download Output"
+            onClick={() => handleDownload(request)}
+          />
+        )}
+      </div>
+      {execute_authority &&
+        requestProjections &&
+        requestProjections.length > 0 && (
+          <div className="card">
+            <h5>Run Next</h5>
+            <Dropdown
+              value={requestProjectionSelected}
+              options={requestProjections}
+              valueTemplate={commandTemplate}
+              itemTemplate={commandTemplate}
+              onChange={(e) => {
+                requestProjectionSelectedRef.current = e.value;
+                setRequestProjectionSelected(e.value);
+              }}
+              placeholder="Select a command to run next"
+            />
+            <AccessButton
+              label="Run"
+              onClick={() => {
+                if (requestProjectionSelectedRef.current) {
+                  addRequestItem({
+                    type: "REQUEST",
+                    requestCommandInput: requestProjectionSelectedRef.current,
+                  });
+                }
+              }}
+            />
+          </div>
+        )}
+    </div>
+  );
+}
+
+export default RequestOptions;
