+++ src/react/src/layouts/RequestView.tsx
@@ -1,212 +1,13 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { BreadCrumb } from "primereact/breadcrumb";
-import { Dropdown } from "primereact/dropdown";
-import { MenuItem } from "primereact/menuitem";
-import { Message } from "primereact/message";
-import { Skeleton } from "primereact/skeleton";
-import { SplitButton } from "primereact/splitbutton";
-import { Stepper } from "primereact/stepper";
-import { StepperPanel } from "primereact/stepperpanel";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 
-import AccessButton from "../components/AccessButton";
-import CommandForm from "../components/CommandForm";
-import RequestOutput from "../components/RequestOutput";
 import RequestTreeChart from "../components/RequestTreeChart";
-import { Request, System } from "../models/brewtils-types";
-import {
-  Config,
-  PermissionCheck,
-  RequestCommand,
-  RequestItem,
-} from "../models/models";
-import { checkPermission } from "../services/permission_service";
-import {
-  CancelRequest,
-  DeleteRequest,
-  GetRequest,
-  GetRequestProjections,
-} from "../services/request_service";
-import { GetSystemList } from "../services/system_service";
-import { GetBaseURL } from "../services/util_service";
-
-function UnformattedInput(request: Request) {
-  return (
-    <div>
-      <Message severity="warn" text="Unable to find source System/Command" />
-      <pre>{JSON.stringify(request.parameters, null, 2)}</pre>
-    </div>
-  );
-}
-
-const handleDownload = (request: Request) => {
-  // Example: fetch a file from a URL
-  const fileUrl = `${GetBaseURL()}/api/v1/requests/output/${request.id}`;
-  let filename = `${request.id}.txt`;
-  if (request.output_type == "HTML") {
-    filename = `${request.id}.html`;
-  } else if (request.output_type == "JSON") {
-    filename = `${request.id}.json`;
-  }
-
-  fetch(fileUrl)
-    .then((response) => response.blob())
-    .then((blob) => {
-      // Create blob link to download
-      const url = window.URL.createObjectURL(new Blob([blob]));
-      const link = document.createElement("a");
-      link.href = url;
-      link.setAttribute("download", filename); // Set the custom download name
-      document.body.appendChild(link);
-      link.click(); // Trigger the download
-      link?.parentNode?.removeChild(link); // Clean up the link
-      window.URL.revokeObjectURL(url); // Free up the memory
-    })
-    .catch((error) => {
-      console.error("Error fetching the file:", error);
-    });
-};
-
-function RequestOptions({
-  request,
-  requestProjections,
-  requestProjectionSelected,
-  setRequestProjectionSelected,
-  requestProjectionSelectedRef,
-  addRequestItem,
-  config,
-}: {
-  request: Request;
-  requestProjections?: RequestCommand[];
-  requestProjectionSelected?: RequestCommand;
-  setRequestProjectionSelected: (value: RequestCommand | undefined) => void;
-  requestProjectionSelectedRef: React.RefObject<RequestCommand | undefined>;
-  addRequestItem: (itemParams?: Partial<RequestItem>) => void;
-  config: Config;
-}) {
-  const navigate = useNavigate();
-  const items: MenuItem[] = [];
-
-  const execute_authority = checkPermission(config, "OPERATOR", {
-    gardenName: request?.target_garden,
-  } as PermissionCheck);
-
-  if (execute_authority) {
-    if (
-      request.status &&
-      ["CREATED", "RECEIVED", "IN_PROGRESS"].includes(request.status)
-    ) {
-      items.push({
-        label: "Cancel Request",
-        icon: <FontAwesomeIcon icon="xmark" />,
-        command: () => {
-          CancelRequest(request).catch((error) => {
-            console.error("Error canceling request:", error);
-          });
-        },
-      });
-    } else {
-      items.push({
-        label: "Download Output",
-        icon: <FontAwesomeIcon icon="download" />,
-        command: () => {
-          handleDownload(request);
-        },
-      });
-
-      if (
-        checkPermission(config, "GARDEN_ADMIN", {
-          gardenName: request?.target_garden,
-        } as PermissionCheck)
-      ) {
-        items.push({
-          label: "Delete Request",
-          icon: <FontAwesomeIcon icon="xmark" />,
-          command: () => {
-            DeleteRequest(request)
-              .then(() => {
-                void navigate(`/requests`);
-              })
-              .catch((error) => {
-                console.error("Error deleting request:", error);
-              });
-          },
-        });
-      }
-    }
-  }
-
-  const pourAgain = (request: Request) => {
-    addRequestItem({ requestId: request.id, type: "REQUEST" });
-  };
-
-  const commandTemplate = (requestCommand: RequestCommand) => {
-    return (
-      <span>
-        {requestCommand?.namespace === request.namespace
-          ? null
-          : `${requestCommand?.namespace} / `}{" "}
-        {requestCommand?.systemName} / {requestCommand?.version} /{" "}
-        {requestCommand?.instance} / {requestCommand?.command}
-      </span>
-    );
-  };
-
-  return (
-    <div className="card justify-content-end">
-      <div className="flex flex-end">
-        {execute_authority && (
-          <SplitButton
-            label="Pour Again"
-            icon={<FontAwesomeIcon icon="plus" />}
-            model={items}
-            className="p-button-secondary"
-            onClick={() => pourAgain(request)}
-            severity="success"
-            style={{ marginLeft: "auto" }}
-          />
-        )}
-        {!execute_authority && (
-          <AccessButton
-            icon={<FontAwesomeIcon icon="download" />}
-            label="Download Output"
-            onClick={() => handleDownload(request)}
-          />
-        )}
-      </div>
-      {execute_authority &&
-        requestProjections &&
-        requestProjections.length > 0 && (
-          <div className="card">
-            <h5>Run Next</h5>
-            <Dropdown
-              value={requestProjectionSelected}
-              options={requestProjections}
-              valueTemplate={commandTemplate}
-              itemTemplate={commandTemplate}
-              onChange={(e) => {
-                requestProjectionSelectedRef.current = e.value;
-                setRequestProjectionSelected(e.value);
-              }}
-              placeholder="Select a command to run next"
-            />
-            <AccessButton
-              label="Run"
-              onClick={() => {
-                if (requestProjectionSelectedRef.current) {
-                  addRequestItem({
-                    type: "REQUEST",
-                    requestCommandInput: requestProjectionSelectedRef.current,
-                  });
-                }
-              }}
-            />
-          </div>
-        )}
-    </div>
-  );
-}
+import RequestViewMain from "../components/RequestViewMain";
+import { Request } from "../models/brewtils-types";
+import { Config, RequestItem } from "../models/models";
+import { GetRequest } from "../services/request_service";
 
 function RequestHeader(request: Request) {
   const iconItemTemplate = (item: any, options: any) => {
@@ -265,21 +66,10 @@ function RequestView({
 }) {
   const { requestId } = useParams<{ requestId: string }>();
   const [request, setRequest] = useState<Request | null>(null);
-  const [system, setSystem] = useState<System | null>(null);
-  const [command, setCommand] = useState<any>(null);
+
   const [rootRequest, setRootRequest] = useState<Request | null>(null);
-  const [showCommandForm, setShowCommandForm] = useState(false);
 
   const rootRequestId = useRef<string | null>(null);
-  const [requestProjections, setRequestProjections] = useState<
-    RequestCommand[] | undefined
-  >(undefined);
-  const [requestProjectionSelected, setRequestProjectionSelected] = useState<
-    RequestCommand | undefined
-  >(undefined);
-  const requestProjectionSelectedRef = useRef<RequestCommand | undefined>(
-    undefined,
-  );
 
   const MonitorRequestId = useCallback(
     (message: any) => {
@@ -325,21 +115,10 @@ function RequestView({
           });
       }
     } else {
-      GetRequestProjections(request)
-        .then((projections) => {
-          setRequestProjections(projections);
-          setRequestProjectionSelected(projections[0]);
-          requestProjectionSelectedRef.current = projections[0];
-        })
-        .catch((error) => {
-          console.error("Error fetching request projections:", error);
-        });
       if (
         request.status &&
         ["CANCELED", "SUCCESS", "ERROR", "INVALID"].includes(request.status)
       ) {
-        setActiveIndex(1);
-
         if (requestId && requestId in listeners) {
           delete listeners[requestId];
         }
@@ -370,32 +149,6 @@ function RequestView({
       };
 
       loadRootRequest(request);
-
-      if (!system) {
-        GetSystemList({
-          name: request.system,
-          version: request.system_version,
-          namespace: request.namespace,
-          garden_name: request.target_garden,
-        })
-          .then((data) => {
-            if (data.length > 0) {
-              setSystem(data[0]);
-            } else {
-              setShowCommandForm(true);
-            }
-          })
-          .catch((error) => {
-            console.error("Error fetching system list:", error);
-            setShowCommandForm(true);
-          });
-      } else if (system.commands) {
-        const commandData = system.commands.find(
-          (cmd) => cmd.name === request.command,
-        );
-        setCommand(commandData);
-        setShowCommandForm(true);
-      }
     }
 
     return () => {
@@ -406,10 +159,7 @@ function RequestView({
         delete listeners[rootRequestId.current];
       }
     };
-  }, [request, requestId, listeners, MonitorRequestId, system]);
-
-  const stepperRef = useRef(null);
-  const [activeIndex, setActiveIndex] = useState(0);
+  }, [request, requestId, listeners, MonitorRequestId]);
 
   return (
     <div>
@@ -426,63 +176,14 @@ function RequestView({
       )}
 
       {request && (
-        <Stepper
-          ref={stepperRef}
-          activeStep={activeIndex}
-          style={{ flexBasis: "50rem" }}
-        >
-          <StepperPanel header="Request Parameters">
-            {/* Need to determine if Read Only can still download values */}
-            <div className="flex">
-              {!showCommandForm && <Skeleton width="100%" height="10rem" />}
-              {showCommandForm && command && (
-                <CommandForm
-                  {...{
-                    command: command,
-                    request: request,
-                    setRequest: setRequest,
-                    resetForm: false,
-                    setResetForm: () => {},
-                    setIsFormValid: () => {},
-                  }}
-                />
-              )}
-              {showCommandForm && !command && <UnformattedInput {...request} />}
-
-              {request && (
-                <div style={{ marginLeft: "auto" }}>
-                  <RequestOptions
-                    request={request}
-                    config={config}
-                    addRequestItem={addRequestItem}
-                    requestProjections={requestProjections}
-                    requestProjectionSelected={requestProjectionSelected}
-                    setRequestProjectionSelected={setRequestProjectionSelected}
-                    requestProjectionSelectedRef={requestProjectionSelectedRef}
-                  />
-                </div>
-              )}
-            </div>
-          </StepperPanel>
-          <StepperPanel header="Request Output">
-            <div className="flex">
-              {request && <RequestOutput {...request} />}
-              {request && (
-                <div style={{ marginLeft: "auto" }}>
-                  <RequestOptions
-                    request={request}
-                    config={config}
-                    addRequestItem={addRequestItem}
-                    requestProjections={requestProjections}
-                    requestProjectionSelected={requestProjectionSelected}
-                    setRequestProjectionSelected={setRequestProjectionSelected}
-                    requestProjectionSelectedRef={requestProjectionSelectedRef}
-                  />
-                </div>
-              )}
-            </div>
-          </StepperPanel>
-        </Stepper>
+        <RequestViewMain
+          request={request}
+          setRequest={setRequest}
+          addRequestItem={addRequestItem}
+          showProjections={true}
+          config={config}
+          isCard={false}
+        />
       )}
     </div>
   );
